Deleted HardwareMap, split HardwareMap into AutoHardwareMap and TeleOpHardwareMap, edited the TeleOpMethods program accordingly, and deleted IBTest.java
