+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/AutoMethods.java
@@ -7,7 +7,6 @@ import com.qualcomm.robotcore.util.Range;
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 import org.firstinspires.ftc.robotcore.external.navigation.AxesOrder;
 import org.firstinspires.ftc.robotcore.external.navigation.AxesReference;
-import org.firstinspires.ftc.teamcode.HardwareMap;
 
 public abstract class AutoMethods extends AutoHardwareMap {
     public void initRobot() {
