--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/IBTest.java
@@ -1,9 +1,0 @@
-package org.firstinspires.ftc.teamcode;
-
-/**
- * Created by ibravo on 8/25/18.
- */
-
-public class IBTest {
-    /* new comment */
-}
