+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TeleOp/TeleOpMethods.java
@@ -2,9 +2,7 @@ package org.firstinspires.ftc.teamcode.TeleOp;
 
 import com.qualcomm.robotcore.util.Range;
 
-import org.firstinspires.ftc.teamcode.HardwareMap;
-
-public abstract class TeleOpMethods extends HardwareMap {
+public abstract class TeleOpMethods extends TeleOpHardwareMap {
     public void move (double myangle, float mypower, float myrot) {
         if (leftFrontDrive != null && leftBackDrive != null && rightFrontDrive != null && rightBackDrive != null) {
             leftFrontDrive.setPower(Range.clip(myrot + (mypower * ((Math.sin((myangle + 135) / 180 * 3.141592)))), -1, 1));
